feat: create CatCard component, adjust styling, move cats.js into constants folder
